Switch to use thread local to make things go faster.
