+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/LogicalPlan.scala
@@ -23,141 +23,15 @@ import org.apache.spark.sql.catalyst.analysis._
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.plans.QueryPlan
 import org.apache.spark.sql.catalyst.plans.logical.statsEstimation.LogicalPlanStats
-import org.apache.spark.sql.catalyst.trees.{CurrentOrigin, TreeNode}
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.util.Utils
-
-
-object LogicalPlan {
-
-  private val resolveOperatorDepth = new ThreadLocal[Int] {
-    override def initialValue(): Int = 0
-  }
-
-  def allowInvokingTransformsInAnalyzer[T](f: => T): T = {
-    resolveOperatorDepth.set(resolveOperatorDepth.get + 1)
-    try f finally {
-      resolveOperatorDepth.set(resolveOperatorDepth.get - 1)
-    }
-  }
-}
 
 
 abstract class LogicalPlan
   extends QueryPlan[LogicalPlan]
+  with AnalysisHelper
   with LogicalPlanStats
   with QueryPlanConstraints
   with Logging {
-
-  private var _analyzed: Boolean = false
-
-  /**
-   * Marks this plan as already analyzed. This should only be called by [[CheckAnalysis]].
-   */
-  private[catalyst] def setAnalyzed(): Unit = { _analyzed = true }
-
-  /**
-   * Returns true if this node and its children have already been gone through analysis and
-   * verification.  Note that this is only an optimization used to avoid analyzing trees that
-   * have already been analyzed, and can be reset by transformations.
-   */
-  def analyzed: Boolean = _analyzed
-
-  /**
-   * Returns a copy of this node where `rule` has been recursively applied first to all of its
-   * children and then itself (post-order, bottom-up). When `rule` does not apply to a given node,
-   * it is left unchanged.  This function is similar to `transformUp`, but skips sub-trees that
-   * have already been marked as analyzed.
-   *
-   * @param rule the function use to transform this nodes children
-   */
-  def resolveOperators(rule: PartialFunction[LogicalPlan, LogicalPlan]): LogicalPlan = {
-    if (!analyzed) {
-      LogicalPlan.allowInvokingTransformsInAnalyzer {
-        val afterRuleOnChildren = mapChildren(_.resolveOperators(rule))
-        if (this fastEquals afterRuleOnChildren) {
-          CurrentOrigin.withOrigin(origin) {
-            rule.applyOrElse(this, identity[LogicalPlan])
-          }
-        } else {
-          CurrentOrigin.withOrigin(origin) {
-            rule.applyOrElse(afterRuleOnChildren, identity[LogicalPlan])
-          }
-        }
-      }
-    } else {
-      this
-    }
-  }
-
-  /** Similar to [[resolveOperators]], but does it top-down. */
-  def resolveOperatorsDown(rule: PartialFunction[LogicalPlan, LogicalPlan]): LogicalPlan = {
-    if (!analyzed) {
-      LogicalPlan.allowInvokingTransformsInAnalyzer {
-        val afterRule = CurrentOrigin.withOrigin(origin) {
-          rule.applyOrElse(this, identity[LogicalPlan])
-        }
-
-        // Check if unchanged and then possibly return old copy to avoid gc churn.
-        if (this fastEquals afterRule) {
-          mapChildren(_.resolveOperatorsDown(rule))
-        } else {
-          afterRule.mapChildren(_.resolveOperatorsDown(rule))
-        }
-      }
-    } else {
-      this
-    }
-  }
-
-  /**
-   * Recursively transforms the expressions of a tree, skipping nodes that have already
-   * been analyzed.
-   */
-  def resolveExpressions(r: PartialFunction[Expression, Expression]): LogicalPlan = {
-    this resolveOperators  {
-      case p => p.transformExpressions(r)
-    }
-  }
-
-  protected def assertNotAnalysisRule(): Unit = {
-    if (Utils.isTesting && LogicalPlan.resolveOperatorDepth.get == 0) {
-      if (Thread.currentThread.getStackTrace.exists(_.getClassName.contains("Analyzer"))) {
-        val e = new RuntimeException("This method should not be called in the analyzer")
-        e.printStackTrace()
-        throw e
-      }
-    }
-  }
-
-  /**
-   * In analyzer, use [[resolveOperatorsDown()]] instead. If this is used in the analyzer,
-   * an exception will be thrown in test mode. It is however OK to call this function within
-   * the scope of a [[resolveOperatorsDown()]] call.
-   * @see [[TreeNode.transformDown()]].
-   */
-  override def transformDown(rule: PartialFunction[LogicalPlan, LogicalPlan]): LogicalPlan = {
-    assertNotAnalysisRule()
-    super.transformDown(rule)
-  }
-
-  /**
-   * Use [[resolveOperators()]] in the analyzer.
-   * @see [[TreeNode.transformUp()]]
-   */
-  override def transformUp(rule: PartialFunction[LogicalPlan, LogicalPlan]): LogicalPlan = {
-    assertNotAnalysisRule()
-    super.transformUp(rule)
-  }
-
-  /**
-   * Use [[resolveExpressions()]] in the analyzer.
-   * @see [[QueryPlan.transformAllExpressions()]]
-   */
-  override def transformAllExpressions(rule: PartialFunction[Expression, Expression]): this.type = {
-    assertNotAnalysisRule()
-    super.transformAllExpressions(rule)
-  }
 
   /** Returns true if this subtree has data from a streaming data source. */
   def isStreaming: Boolean = children.exists(_.isStreaming == true)
